chore: fix all clippy lints; clippy -D warnings gates CI
Addresses the newer-toolchain lints properly instead of softening the
gate: darray enumerate + needless return, manual div_ceil in offsets,
redundant cast, as_bytes-after-slice in the CLI dump, auto-deref in
sshash-py, hex-literal grouping in the prefilter test, and every
in-workspace use of the deprecated canonical field/method (tests build
with the default config now; stale canonical-mode doc comments in
minimizer_tuples updated to describe the unified scheme). The only
remaining note is cargo's future-incompat report for proc-macro-error2
(transitive via sux -> impl-tools; upstream).

+++ crates/sshash-lib/src/minimizer.rs
@@ -455,7 +455,7 @@ mod tests {
         let acgt = {
             let mut bits = 0u64;
             for (i, b) in "ACGT".bytes().enumerate() {
-                bits |= ((((b >> 1) & 3) as u64) << (i * 2)) as u64;
+                bits |= (((b >> 1) & 3) as u64) << (i * 2);
             }
             bits
         };
